+++ examples/ccall.rs
@@ -45,6 +45,8 @@ mod tests {
 
     #[test]
     fn call_add() {
+        let id = std::thread::current().id();
+        println!("ID: {:?}", id);
         JULIA.with(|j| {
             let mut jlrs = j.borrow_mut();
             let mut frame = StackFrame::new();
@@ -71,6 +73,8 @@ mod tests {
 
     #[test]
     fn call_incr_array() {
+        let id = std::thread::current().id();
+        println!("ID: {:?}", id);
         JULIA.with(|j| {
             let mut jlrs = j.borrow_mut();
             let mut frame = StackFrame::new();