+++ examples/ccall_with_threads.rs
@@ -8,7 +8,7 @@ use std::{
 use thread::JoinHandle;
 
 // This crate is called `ccall_with_threads`, so the library is called
-// `libccall_with_threads`. The functions are  annotated with `no_mangle` to prevent name mangling
+// `libccall_with_threads`. The functions are annotated with `no_mangle` to prevent name mangling
 // and `extern "C"` to make them callable with the C ABI.
 
 /// The handle field of an `AsyncCondition`.